refactor: move CellExtractor to JsonObjectCellExtractor and extract JsonByteExtractor
